requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
reportlab
matplotlib
wordcloud
plotly